etl_pokemon/requirements.txt
# ETL Pokemon Requirements
# Updated: 2026-01-29
# Versions aligned with Docker production environment

# --- Core ---
python-dotenv==1.2.1

# --- API ---
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5

# --- Database ---
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
asyncpg==0.31.0

# --- HTTP / Async ---
requests==2.32.5
aiohttp==3.11.11

# --- Data ---
pandas==2.3.3
numpy==2.4.1

# --- Scraping ---
scrapy==2.12.0
parsel==1.9.1
lxml==5.3.0

# --- Utils ---
tqdm==4.67.1
